+++ packages/ruleset-migrator/src/__tests__/__fixtures__/aliases-variant-1/output.cjs
@@ -1,7 +1,7 @@
 module.exports = {
   aliases: {
-    PathItem: '$.paths[*][*]',
-    Description: '$..description',
-    Name: '$..name',
+    PathItem: ['$.paths[*][*]'],
+    Description: ['$..description'],
+    Name: ['$..name'],
   },
 };